remove: 중복된 기능 폴더 삭제
