+++ design-system/ui/texts/DayCounter.tsx
@@ -1,18 +1,27 @@
 import React from 'react';
+import { flexCenter } from '../../styles/flex';
 
-interface DayCounterProps {
-  children: string;
-  isChecked: boolean;
-  className?: React.ReactNode;
+interface CountdownProps {
+  children: React.ReactNode; // 버튼 내부 컨텐츠
+  isChecked: boolean; // 활성화 여부
 }
 
-const DayCounter = ({ children, isChecked, className }: DayCounterProps) => {
-  const activeStyles = `border-deDayText text-deDayText bg-deDayBg`;
-  const inactiveStyles = `border-deDayTextDark text-deDayTextDark bg-deDayBgLight`;
-  const baseStyles = `h-6 sm:h-4 md:h-6 lg:h-6 px-2 rounded-[2px] text-xs sm:text-xs md:text-xs lg:text-sm 
-    border-[0.1px] font-medium`;
+const Countdown = ({ children, isChecked }: CountdownProps) => {
+  // Active와 Inactive 스타일 분기
+  const activeStyles = `
+    border-deDayText text-deDayText bg-deDayBg
+  `;
 
-  return <div className={`${className} ${baseStyles} ${isChecked ? activeStyles : inactiveStyles}`}>{children}</div>;
+  const inactiveStyles = `
+    border-deDayTextDark text-deDayTextDark bg-deDayBgLight
+  `;
+
+  const baseStyles = `
+    h-5 sm:h-4 md:h-5 px-2 py-1 rounded-[2px] text-xs 
+    border-[0.1px] font-medium ${flexCenter} min-w-[38px] max-w-[38px]
+  `;
+
+  return <button className={`${baseStyles} ${isChecked ? activeStyles : inactiveStyles}`}>{children}</button>;
 };
 
-export default DayCounter;
+export default Countdown;
